+++ qa/workunits/windows/run-tests.ps1
@@ -4,7 +4,7 @@ $ErrorActionPreference = "Stop"
 $scriptLocation = [System.IO.Path]::GetDirectoryName(
     $myInvocation.MyCommand.Definition)
 
-$testRbdWnbd = "$scriptLocation/test_rbd_wnbd.py"
+$env:PYTHONPATH += ";$scriptLocation"
 
 function safe_exec() {
     # Powershell doesn't check the command exit code, we'll need to
@@ -16,14 +16,13 @@ function safe_exec() {
     }
 }
 
-safe_exec python.exe $testRbdWnbd --test-name RbdTest --iterations 100
-safe_exec python.exe $testRbdWnbd --test-name RbdFioTest --iterations 100
-safe_exec python.exe $testRbdWnbd --test-name RbdStampTest --iterations 100
+safe_exec python.exe -m py_tests.rbd_wnbd.stress_test --test-name RbdTest --iterations 100
+safe_exec python.exe -m py_tests.rbd_wnbd.stress_test --test-name RbdFioTest --iterations 100
+safe_exec python.exe -m py_tests.rbd_wnbd.stress_test --test-name RbdStampTest --iterations 100
 
 # It can take a while to setup the partition (~10s), we'll use fewer iterations.
-safe_exec python.exe $testRbdWnbd --test-name RbdFsTest --iterations 4
-safe_exec python.exe $testRbdWnbd --test-name RbdFsFioTest --iterations 4
-safe_exec python.exe $testRbdWnbd --test-name RbdFsStampTest --iterations 4
+safe_exec python.exe -m py_tests.rbd_wnbd.stress_test --test-name RbdFsTest --iterations 4
+safe_exec python.exe -m py_tests.rbd_wnbd.stress_test --test-name RbdFsFioTest --iterations 4
+safe_exec python.exe -m py_tests.rbd_wnbd.stress_test --test-name RbdFsStampTest --iterations 4
 
-safe_exec python.exe $testRbdWnbd `
-    --test-name RbdResizeFioTest --image-size-mb 64
+safe_exec python.exe -m py_tests.rbd_wnbd.stress_test --test-name RbdResizeFioTest --image-size-mb 64
